+++ include/rock/tree_set.h
@@ -42,7 +42,8 @@ bool rock_tree_set_init(struct rock_tree_set *object,
  * @brief Invalidate tree set.
  * <p>All the items contained within the tree set will have the given <i>on
  * destroy</i> callback invoked upon itself. The actual <u>tree set instance
- * is not deallocated</u> since it may have been embedded in a larger structure.
+ * is not deallocated</u> since it may have been embedded in a larger
+ * structure.</p>
  * @param [in] object instance to be invalidated.
  * @param [in] on_destroy called just before the item is to be destroyed.
  * @return On success true, otherwise false if an error has occurred.
@@ -59,7 +60,7 @@ bool rock_tree_set_invalidate(struct rock_tree_set *object,
  * @throws ROCK_TREE_SET_ERROR_OBJECT_IS_NULL if object is <i>NULL</i>.
  * @throws ROCK_TREE_SET_ERROR_OUT_IS_NULL if out is <i>NULL</i>.
  */
-bool rock_tree_set_count(struct rock_tree_set *object, size_t *out);
+bool rock_tree_set_count(struct rock_tree_set *object, uintmax_t *out);
 
 /**
  * @brief Add item to the tree set.
