+++ src/error.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
 #include <rock.h>
 
-_Thread_local size_t rock_error = ROCK_ERROR_NONE;
+_Thread_local uintmax_t rock_error = ROCK_ERROR_NONE;
